+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0 - (2022-03-23)
+
+* Type/value-safe layer to adapt to server-defined endpoints.
+* Request arguments checks to capture errors in request construction.
+
 ## 0.1.4 - (2022-03-11)
 
 * Bug fix (aiohttp session was not being closed in HTTPClientJWTRenewable)